requirements.txt
feedparser==6.0.10
requests==2.31.0
gradio==4.19.2
typing-extensions==4.9.0
openai==1.12.0
python-dotenv==1.0.1